requirements.txt
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
newsapi-python>=0.2.7

